+++ contracts/test/v0.8/automation/AutomationRegistrar2_1.test.ts
@@ -12,7 +12,7 @@ import { UpkeepMock } from '../../../typechain/UpkeepMock'
 import { toWei } from '../../test-helpers/helpers'
 import { IKeeperRegistryMaster as IKeeperRegistry } from '../../../typechain/IKeeperRegistryMaster'
 import { AutomationRegistrar2_1 as Registrar } from '../../../typechain/AutomationRegistrar2_1'
-import { deployRegistry21, encodeConfig21 } from './helpers'
+import { deployRegistry21 } from './helpers'
 
 // copied from KeeperRegistryBase2_1.sol
 enum Trigger {
@@ -48,7 +48,7 @@ describe('AutomationRegistrar2_1', () => {
 
   const linkEth = BigNumber.from(300000000)
   const gasWei = BigNumber.from(100)
-  const executeGas = BigNumber.from(100000)
+  const performGas = BigNumber.from(100000)
   const paymentPremiumPPB = BigNumber.from(250000000)
   const flatFeeMicroLink = BigNumber.from(0)
   const maxAllowedAutoApprove = 5
@@ -144,7 +144,7 @@ describe('AutomationRegistrar2_1', () => {
       await personas.Ned.getAddress(),
       await personas.Neil.getAddress(),
     ]
-    const onchainConfig = encodeConfig21({
+    const onchainConfig = {
       paymentPremiumPPB,
       flatFeeMicroLink,
       checkGasLimit,
@@ -160,10 +160,12 @@ describe('AutomationRegistrar2_1', () => {
       transcoder,
       registrars: [registrar.address],
       upkeepPrivilegeManager: upkeepManager,
-    })
+    }
     await registry
       .connect(owner)
-      .setConfig(keepers, keepers, 1, onchainConfig, 1, '0x')
+      [
+        'setConfig(address[],address[],uint8,(uint32,uint32,uint32,uint24,uint16,uint96,uint32,uint32,uint32,uint32,uint256,uint256,address,address[],address),uint64,bytes)'
+      ](keepers, keepers, 1, onchainConfig, 1, '0x')
   })
 
   describe('#typeAndVersion', () => {
@@ -182,7 +184,7 @@ describe('AutomationRegistrar2_1', () => {
             upkeepName,
             emptyBytes,
             mock.address,
-            executeGas,
+            performGas,
             await admin.getAddress(),
             0,
             emptyBytes,
@@ -210,7 +212,7 @@ describe('AutomationRegistrar2_1', () => {
           upkeepName,
           emptyBytes,
           mock.address,
-          executeGas,
+          performGas,
           await admin.getAddress(),
           0,
           emptyBytes,
@@ -236,7 +238,7 @@ describe('AutomationRegistrar2_1', () => {
           upkeepName,
           emptyBytes,
           mock.address,
-          executeGas,
+          performGas,
           await admin.getAddress(),
           0,
           emptyBytes,
@@ -261,7 +263,7 @@ describe('AutomationRegistrar2_1', () => {
           upkeepName,
           emptyBytes,
           mock.address,
-          executeGas,
+          performGas,
           '0x0000000000000000000000000000000000000000',
           0,
           emptyBytes,
@@ -297,7 +299,7 @@ describe('AutomationRegistrar2_1', () => {
           upkeepName,
           emptyBytes,
           mock.address,
-          executeGas,
+          performGas,
           await admin.getAddress(),
           0,
           emptyBytes,
@@ -319,7 +321,7 @@ describe('AutomationRegistrar2_1', () => {
       assert.equal(newupkeep.admin, await admin.getAddress())
       assert.equal(newupkeep.checkData, emptyBytes)
       assert.equal(newupkeep.balance.toString(), amount.toString())
-      assert.equal(newupkeep.executeGas, executeGas.toNumber())
+      assert.equal(newupkeep.performGas, performGas.toNumber())
       assert.equal(newupkeep.offchainConfig, offchainConfig)
 
       await expect(tx).to.emit(registrar, 'RegistrationRequested')
@@ -346,7 +348,7 @@ describe('AutomationRegistrar2_1', () => {
           upkeepName,
           emptyBytes,
           mock.address,
-          executeGas,
+          performGas,
           await admin.getAddress(),
           0,
           emptyBytes,
@@ -394,7 +396,7 @@ describe('AutomationRegistrar2_1', () => {
         upkeepName,
         emptyBytes,
         mock.address,
-        executeGas,
+        performGas,
         await admin.getAddress(),
         0,
         emptyBytes,
@@ -413,7 +415,7 @@ describe('AutomationRegistrar2_1', () => {
         upkeepName,
         emptyBytes,
         mock.address,
-        executeGas.toNumber() + 1, // make unique hash
+        performGas.toNumber() + 1, // make unique hash
         await admin.getAddress(),
         0,
         emptyBytes,
@@ -432,7 +434,7 @@ describe('AutomationRegistrar2_1', () => {
         upkeepName,
         emptyBytes,
         mock.address,
-        executeGas,
+        performGas,
         await admin.getAddress(),
         Trigger.LOG,
         emptyBytes,
@@ -455,7 +457,7 @@ describe('AutomationRegistrar2_1', () => {
         upkeepName,
         emptyBytes,
         mock.address,
-        executeGas.toNumber() + 2, // make unique hash
+        performGas.toNumber() + 2, // make unique hash
         await admin.getAddress(),
         0,
         emptyBytes,
@@ -474,7 +476,7 @@ describe('AutomationRegistrar2_1', () => {
         upkeepName,
         emptyBytes,
         mock.address,
-        executeGas.toNumber() + 3, // make unique hash
+        performGas.toNumber() + 3, // make unique hash
         await admin.getAddress(),
         0,
         emptyBytes,
@@ -513,7 +515,7 @@ describe('AutomationRegistrar2_1', () => {
           upkeepName,
           emptyBytes,
           mock.address,
-          executeGas,
+          performGas,
           await admin.getAddress(),
           0,
           emptyBytes,
@@ -535,7 +537,7 @@ describe('AutomationRegistrar2_1', () => {
       assert.equal(newupkeep.admin, await admin.getAddress())
       assert.equal(newupkeep.checkData, emptyBytes)
       assert.equal(newupkeep.balance.toString(), amount.toString())
-      assert.equal(newupkeep.executeGas, executeGas.toNumber())
+      assert.equal(newupkeep.performGas, performGas.toNumber())
 
       await expect(tx).to.emit(registrar, 'RegistrationRequested')
       await expect(tx).to.emit(registrar, 'RegistrationApproved')
@@ -566,7 +568,7 @@ describe('AutomationRegistrar2_1', () => {
           upkeepName,
           emptyBytes,
           mock.address,
-          executeGas,
+          performGas,
           await admin.getAddress(),
           0,
           emptyBytes,
@@ -603,7 +605,7 @@ describe('AutomationRegistrar2_1', () => {
         registrar.connect(someAddress).registerUpkeep({
           name: upkeepName,
           upkeepContract: mock.address,
-          gasLimit: executeGas,
+          gasLimit: performGas,
           adminAddress: await admin.getAddress(),
           triggerType: 0,
           checkData: emptyBytes,
@@ -632,7 +634,7 @@ describe('AutomationRegistrar2_1', () => {
         registrar.connect(someAddress).registerUpkeep({
           name: upkeepName,
           upkeepContract: mock.address,
-          gasLimit: executeGas,
+          gasLimit: performGas,
           adminAddress: await admin.getAddress(),
           triggerType: 0,
           checkData: emptyBytes,
@@ -660,7 +662,7 @@ describe('AutomationRegistrar2_1', () => {
       const tx = await registrar.connect(requestSender).registerUpkeep({
         name: upkeepName,
         upkeepContract: mock.address,
-        gasLimit: executeGas,
+        gasLimit: performGas,
         adminAddress: await admin.getAddress(),
         triggerType: 0,
         checkData: emptyBytes,
@@ -678,7 +680,7 @@ describe('AutomationRegistrar2_1', () => {
       assert.equal(newupkeep.admin, await admin.getAddress())
       assert.equal(newupkeep.checkData, emptyBytes)
       assert.equal(newupkeep.balance.toString(), amount.toString())
-      assert.equal(newupkeep.executeGas, executeGas.toNumber())
+      assert.equal(newupkeep.performGas, performGas.toNumber())
       assert.equal(newupkeep.offchainConfig, offchainConfig)
 
       await expect(tx).to.emit(registrar, 'RegistrationRequested')
@@ -760,7 +762,7 @@ describe('AutomationRegistrar2_1', () => {
           upkeepName,
           emptyBytes,
           mock.address,
-          executeGas,
+          performGas,
           await admin.getAddress(),
           0,
           emptyBytes,
@@ -784,7 +786,7 @@ describe('AutomationRegistrar2_1', () => {
         .approve(
           upkeepName,
           mock.address,
-          executeGas,
+          performGas,
           await admin.getAddress(),
           0,
           emptyBytes,
@@ -801,7 +803,7 @@ describe('AutomationRegistrar2_1', () => {
         .approve(
           upkeepName,
           mock.address,
-          executeGas,
+          performGas,
           await admin.getAddress(),
           0,
           emptyBytes,
@@ -818,7 +820,7 @@ describe('AutomationRegistrar2_1', () => {
         .approve(
           upkeepName,
           ethers.Wallet.createRandom().address,
-          executeGas,
+          performGas,
           await admin.getAddress(),
           0,
           emptyBytes,
@@ -846,7 +848,7 @@ describe('AutomationRegistrar2_1', () => {
         .approve(
           upkeepName,
           mock.address,
-          executeGas,
+          performGas,
           ethers.Wallet.createRandom().address,
           0,
           emptyBytes,
@@ -860,7 +862,7 @@ describe('AutomationRegistrar2_1', () => {
         .approve(
           upkeepName,
           mock.address,
-          executeGas,
+          performGas,
           await admin.getAddress(),
           0,
           '0x1234',
@@ -877,7 +879,7 @@ describe('AutomationRegistrar2_1', () => {
         .approve(
           upkeepName,
           mock.address,
-          executeGas,
+          performGas,
           await admin.getAddress(),
           0,
           emptyBytes,
@@ -894,7 +896,7 @@ describe('AutomationRegistrar2_1', () => {
         .approve(
           upkeepName,
           mock.address,
-          executeGas,
+          performGas,
           await admin.getAddress(),
           0,
           emptyBytes,
@@ -907,7 +909,7 @@ describe('AutomationRegistrar2_1', () => {
         .approve(
           upkeepName,
           mock.address,
-          executeGas,
+          performGas,
           await admin.getAddress(),
           0,
           emptyBytes,
@@ -938,7 +940,7 @@ describe('AutomationRegistrar2_1', () => {
           upkeepName,
           emptyBytes,
           mock.address,
-          executeGas,
+          performGas,
           await admin.getAddress(),
           0,
           emptyBytes,
@@ -998,7 +1000,7 @@ describe('AutomationRegistrar2_1', () => {
         .approve(
           upkeepName,
           mock.address,
-          executeGas,
+          performGas,
           await admin.getAddress(),
           0,
           emptyBytes,