+++ core/services/gateway/handlers/common/requestcache.go
@@ -24,9 +24,10 @@ type RequestCache[T any] interface {
 type ResponseProcessor[T any] func(response *api.Message, state *T) (aggregated *handlers.UserCallbackPayload, newState *T, err error)
 
 type requestCache[T any] struct {
-	cache   map[globalId]*pendingRequest[T]
-	timeout time.Duration
-	mu      sync.Mutex
+	cache        map[globalId]*pendingRequest[T]
+	maxCacheSize uint32
+	timeout      time.Duration
+	mu           sync.Mutex
 }
 
 type globalId struct {
@@ -41,8 +42,8 @@ type pendingRequest[T any] struct {
 	mu           sync.Mutex
 }
 
-func NewRequestCache[T any](timeout time.Duration) RequestCache[T] {
-	return &requestCache[T]{cache: make(map[globalId]*pendingRequest[T]), timeout: timeout}
+func NewRequestCache[T any](timeout time.Duration, maxCacheSize uint32) RequestCache[T] {
+	return &requestCache[T]{cache: make(map[globalId]*pendingRequest[T]), timeout: timeout, maxCacheSize: maxCacheSize}
 }
 
 func (c *requestCache[T]) NewRequest(request *api.Message, callbackCh chan<- handlers.UserCallbackPayload, responseData *T) error {
@@ -59,8 +60,11 @@ func (c *requestCache[T]) NewRequest(request *api.Message, callbackCh chan<- han
 	if ok {
 		return errors.New("request already exists")
 	}
+	if len(c.cache) >= int(c.maxCacheSize) {
+		return errors.New("request cache is full")
+	}
 	timer := time.AfterFunc(c.timeout, func() {
-		c.deleteAndSendOnce(key, handlers.UserCallbackPayload{Msg: request, ErrCode: api.RequestTimeoutError})
+		c.deleteAndSendOnce(key, handlers.UserCallbackPayload{Msg: request, ErrMsg: "timeout", ErrCode: api.RequestTimeoutError})
 	})
 	c.cache[key] = &pendingRequest[T]{callbackCh: callbackCh, responseData: responseData, timeoutTimer: timer}
 	return nil