+++ docs/CHANGELOG.md
@@ -9,7 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [dev]
 
-...
+### Fixed
+- Updated `v2/keys/evm` and `v2/keys/eth` routes to return 400 and 404 status codes where appropriate. Previously 500s were returned when requested resources were not found or client requests could not be parsed. 
+
+### Added
+- Added the ability to specify and merge fields from multiple secrets files. Overrides of fields and keys are not allowed.
 
 ## 2.3.0 - UNRELEASED